# [PERF] account: speedup bank_statement_line creation