When importing statements the creation of the `account_move_lines` in `account_bank_statement_line.create` takes a lot of time. One of the reasons for that is that the amls are created for each st_line via the one2many field assignation.

This commit speeds up this process by creating all the amls at once after looping through `st_lines`. Batching the amls creation like that leads to a noticeable speedup when importing a lot of statements.

speedup

Importing bank_statement file in Odoo 18.0.

| Number of lines_to_create | Before PR  | After PR |
|:-------------------------:|:----------:|:--------:|
|           100             |    5.24s   |  2.78s   |
|           500             |    22.72s  |  11.47s  |
|           1000            |    44.66s  |  21.73s  |
|           3000            |    2min15s |  1min05s |


---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr