@robodoo 2025-03-27
[![Pull request status dashboard](https://mergebot.odoo.com/odoo/odoo/pull/203681.png)](https://mergebot.odoo.com/odoo/odoo/pull/203681)